[FIX] project_forecast_line: error in consolidated capacity
During some operations, forecast lines are deleted and recreated. This
can lead to some hr.employee.forecast.role lines being deleted which
leaves the forecast lines on the same period without a related document
to store the consolidated capacity. Normally that line is recreated
shortly afterwards but the creation does not recompute the link between
the parent-less lines and the new one. This patches forces the
recomputation when new lines are created.