[FIX] project_forecast_line: disappearing forecast
Fix an issue in project_forecast_line where lines would be removed by
the cron.

When an employee has a forecast role with an end date set, depending on
the order of the processing of records in the cron, a bug could remove
forecast lines of other employees when attempting to remove the forecast
lines related to thatrole which had been generated after the end date in
previous runs (before the end date was set).

Fix this by using the correct domain.